requirements.txt
fastapi
uvicorn[standard]
azure-cosmos
azure-identity
python-jose[cryptography]
pytest
unittest
